requirements.txt
asn1crypto==0.24.0
backports-abc==0.5
certifi==2020.6.20
cffi==1.14.1
cryptography==3.0
decorator==4.4.2
duplicity==0.7.17
enum34==1.1.10
escape==1.1
fasteners==0.12.0
futures==3.3.0
html==1.16
idna==2.6
ipaddress==1.0.23
keyring==10.6.0
keyrings.alt==3.0
lockfile==0.12.2
monotonic==1.0
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
python-telegram-bot==12.8
pyxdg==0.25
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.15.0
supervisor==4.1.0
tornado==5.1.1
